requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==2.0.15
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
